Added studentDTO class

+++ src/main/kotlin/ua/edu/ukma/dudes/scheduleMeBaby/repository/StudentRepository.kt
@@ -3,4 +3,4 @@ package ua.edu.ukma.dudes.scheduleMeBaby.repository
 import org.springframework.data.repository.CrudRepository
 import ua.edu.ukma.dudes.scheduleMeBaby.entity.Student
 
-interface StudentRepository : CrudRepository<Student, Int>
+interface StudentRepository : CrudRepository<Student, Long>
